+++ src/main/java/com/aetherteam/aether/attachment/AetherPlayerAttachment.java
@@ -216,7 +216,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
     public void onJoinLevel(Player player) {
         if (player.level().isClientSide() && player.isLocalPlayer()) {
             CustomizationsOptions.INSTANCE.load();
-            this.setSynced(player.getId(), SyncTarget.SERVER, SHOULD_SYNC_BETWEEN_CLIENTS_SYNC_KEY, true);
+            this.setSyncedToServer(player.getId(), SHOULD_SYNC_BETWEEN_CLIENTS_SYNC_KEY, true);
         }
     }
 
@@ -258,7 +258,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
 
     private void syncAfterJoin(Player player) {
         if (this.shouldSyncAfterJoin) {
-            this.forceSync(player.getId(), AttachmentSyncable.SyncTarget.CLIENT);
+            this.forceSyncToClients(player.getId());
             this.shouldSyncAfterJoin = false;
         }
     }
@@ -271,7 +271,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
                     PlayerList playerList = server.getPlayerList();
                     for (ServerPlayer serverPlayer : playerList.getPlayers()) {
                         if (!serverPlayer.getUUID().equals(player.getUUID())) {
-                            player.getAttachedOrCreate(AetherDataAttachments.AETHER_PLAYER).forceSync(player.getId(), AttachmentSyncable.SyncTarget.CLIENT);
+                            player.getAttachedOrCreate(AetherDataAttachments.AETHER_PLAYER).forceSyncToClients(player.getId());
                         }
                     }
                 }
@@ -363,7 +363,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
 
                 --this.removeGoldenDartTime;
                 if (this.removeGoldenDartTime <= 0) {
-                    this.setSynced(player.getId(), AttachmentSyncable.SyncTarget.CLIENT, GOLDEN_DART_COUNT_SYNC_KEY, this.getGoldenDartCount() - 1);
+                    this.setSyncedToClients(player.getId(), GOLDEN_DART_COUNT_SYNC_KEY, this.getGoldenDartCount() - 1);
                 }
             }
             if (this.getPoisonDartCount() > 0) {
@@ -373,7 +373,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
 
                 --this.removePoisonDartTime;
                 if (this.removePoisonDartTime <= 0) {
-                    this.setSynced(player.getId(), AttachmentSyncable.SyncTarget.CLIENT, POISON_DART_COUNT_SYNC_KEY, this.getPoisonDartCount() - 1);
+                    this.setSyncedToClients(player.getId(), POISON_DART_COUNT_SYNC_KEY, this.getPoisonDartCount() - 1);
                 }
             }
             if (this.getEnchantedDartCount() > 0) {
@@ -383,7 +383,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
 
                 --this.removeEnchantedDartTime;
                 if (this.removeEnchantedDartTime <= 0) {
-                    this.setSynced(player.getId(), AttachmentSyncable.SyncTarget.CLIENT, ENCHANTED_DART_COUNT_SYNC_KEY, this.getEnchantedDartCount() - 1);
+                    this.setSyncedToClients(player.getId(), ENCHANTED_DART_COUNT_SYNC_KEY, this.getEnchantedDartCount() - 1);
                 }
             }
         }
@@ -433,7 +433,7 @@ public class AetherPlayerAttachment implements AttachmentSyncable {
             if (this.attackedWithInvisibility()) {
                 --this.invisibilityAttackCooldown;
                 if (this.invisibilityAttackCooldown <= 0) {
-                    this.setSynced(player.getId(), AttachmentSyncable.SyncTarget.CLIENT, ATTACKED_WITH_INVISIBILITY_SYNC_KEY, false);
+                    this.setSyncedToClients(player.getId(), ATTACKED_WITH_INVISIBILITY_SYNC_KEY, false);
                 }
             } else {
                 this.invisibilityAttackCooldown = AetherConfig.SERVER.invisibility_visibility_time.get();