+++ src/main/java/com/aetherteam/aether/command/EternalDayCommand.java
@@ -3,7 +3,6 @@ package com.aetherteam.aether.command;
 import com.aetherteam.aether.AetherConfig;
 import com.aetherteam.aether.attachment.AetherDataAttachments;
 import com.aetherteam.aether.attachment.AetherTimeAttachment;
-import com.aetherteam.aether.attachment.AttachmentSyncable;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.arguments.BoolArgumentType;
 import net.minecraft.commands.CommandSourceStack;
@@ -33,7 +32,7 @@ public class EternalDayCommand {
             data.setEternalDay(value);
             data.updateEternalDay(level); // Syncs to client.
             if (AetherConfig.SERVER.sync_aether_time.get()) {
-                data.setSynced(-1, AttachmentSyncable.SyncTarget.DIMENSION, AetherTimeAttachment.SHOULD_WAIT_SYNC_KEY, true, level);
+                data.setSyncedToDimension(-1, AetherTimeAttachment.SHOULD_WAIT_SYNC_KEY, true, level);
             }
             source.sendSuccess(() -> Component.translatable("commands.aether.capability.time.eternal_day.set", value), true);
         }