+++ src/main/java/com/aetherteam/aether/network/packet/clientbound/HealthResetPacket.java
@@ -3,7 +3,6 @@ package com.aetherteam.aether.network.packet.clientbound;
 import com.aetherteam.aether.Aether;
 import com.aetherteam.aether.attachment.AetherDataAttachments;
 import com.aetherteam.aether.attachment.AetherPlayerAttachment;
-import com.aetherteam.aether.attachment.AttachmentSyncable;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
@@ -36,7 +35,7 @@ public record HealthResetPacket(int entityID, int value) implements CustomPacket
     public static void execute(HealthResetPacket payload, @Nullable Player contextPlayer) {
         if (contextPlayer != null && contextPlayer.level().getEntity(payload.entityID()) instanceof Player player) {
             var data = player.getAttachedOrCreate(AetherDataAttachments.AETHER_PLAYER);
-            data.setSynced(player.getId(), AttachmentSyncable.SyncTarget.SERVER, AetherPlayerAttachment.LIFE_SHARD_COUNT_SYNC_KEY, payload.value());
+            data.setSyncedToServer(player.getId(), AetherPlayerAttachment.LIFE_SHARD_COUNT_SYNC_KEY, payload.value());
             AttributeInstance attribute = player.getAttribute(Attributes.MAX_HEALTH);
             if (attribute != null) {
                 attribute.removeModifier(data.getLifeShardHealthAttributeModifier().id());